pyproject.toml
[project]
name = "gitinterrogator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain>=1.2.10",
    "langchain-community>=0.4.1",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.10",
    "pygithub>=2.8.1",
    "python-dotenv>=1.2.2",
    "qdrant-client>=1.17.0",
    "sentence-transformers>=5.2.3",
    "tiktoken>=0.12.0",
]
